LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/data-iframe-expected.txt



PASS navigator.serviceWorker is not available in a data: iframe 
